renamed more files and updated their imports
