+++ config.py
@@ -8,6 +8,6 @@ USP_PATH = BASE_PATH / "db" / "procedures"
 # healthDB_DEV.db = DEV
 ENV_FLAG = 0 # 0 = dev, 1 = prod
 if ENV_FLAG == 0:
-    db_name = 'healthDB_DEV.db'
+    DB_NAME = 'healthDB_DEV.db'
 elif ENV_FLAG == 1:
-    db_name = 'healthDB.db'
+    DB_NAME = 'healthDB.db'
